Add support for finding words not found in the original OCR results
The program is now capable of finding words that were not initially found in the OCR search results of the given puzzle image. This is done by searching for every possible orientation of a word, comparing their match values (i.e. a 42% match vs a 96% match) and only updating the result image with the highest matching orientation.

I do realise that this does (relatively) significantly increases the runtime of the program, but for the moment, I think that not performing this search and only matching ~13/40 results vs matching 40/40 results is a worthwhile trade off. For those curious, the time differences on my machine differed by around 1 minute (~10 seconds --> ~70 seconds).