Separate functions into separate files, Remove select print statements
This goes along with the principle of high-cohesion, low-coupling in that related functions are grouped together in the same file and rely minimally on other files to function.

Removed select print statements which aided in development, but are unnecessary for most users. I chose to leave in those which conveyed progress to the user in the event that the program ran for a long period of time so that the user would not think the program crashed.